a* weights added

+++ Assets/Scripts/Tile/Tile.cs
@@ -8,11 +8,11 @@ namespace ThetaStar.Grid
     {
         public Vector3 Position;
         public Vector3 PositionTopLeftCorner;
-        public short Weight;
+        public float Weight;
         public int RowIdx;
         public int ColIdx;
 
-        public Tile(Vector3 position, Vector3 positionTopLeftCorner, short weight, int row, int col)
+        public Tile(Vector3 position, Vector3 positionTopLeftCorner, float weight, int row, int col)
         {
             Position = position;
             PositionTopLeftCorner = positionTopLeftCorner;
